fixed few issues with the RBDI requests and added RBDI PIDs

+++ sw/firmware/main/include/webserver.hpp
@@ -30,5 +30,5 @@ esp_err_t p_vin_index_handler(httpd_req_t *req, void *arg);
 esp_err_t p_clear_dtc_index_handler(httpd_req_t *req, void *arg);
 esp_err_t ws_socket_handler(httpd_req_t *req);
 
-void pid_stream_callback(uint8_t pid);
+void pid_stream_callback(uint16_t pid);
 void enable_pid_stream(bool enable);
